requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0

# LLM API clients
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.28.0

# Data visualization
plotly>=5.15.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Document generation
python-docx>=0.8.11

# Data processing & analysis
datasets>=2.14.0
pathlib>=1.0.1
re>=2.2.1

# Environment & configuration
python-dotenv>=1.0.0

# Additional packages used in the project
json5>=0.9.0
typing-extensions>=4.5.0

# Note: pyarrow auto-installed with pandas/streamlit
# If build issues occur on Windows Python 3.14+:
# pip install --only-binary=all pyarrow

# Optional dependencies for analysis scripts
# Uncomment if needed:
# nltk>=3.8.0
# scipy>=1.10.0
# scikit-learn>=1.3.0

# Development/testing dependencies
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0